@@ -2,7 +2,7 @@
 # Module
 #     setup.cfg
 # Copyright
-#     Copyright (C) 2017 Vladimir Roncevic <elektron.ronca@gmail.com>
+#     Copyright (C) 2017 - 2024 Vladimir Roncevic <elektron.ronca@gmail.com>
 #     gen_unnamed_pipe is free software: you can redistribute it and/or modify
 #     it under the terms of the GNU General Public License as published by the
 #     Free Software Foundation, either version 3 of the License, or
@@ -14,33 +14,24 @@
 #     You should have received a copy of the GNU General Public License along
 #     with this program. If not, see <http://www.gnu.org/licenses/>.
 # Info
-#     Defined setup for package gen_unnamed_pipe.
+#     Defines setup for package gen_unnamed_pipe.
 #
 
 [metadata]
 name = gen_unnamed_pipe
-version = 1.0.3
+version = 1.0.4
 author = Vladimir Roncevic
 author_email = elektron.ronca@gmail.com
-description = Unnamed PIPE Code Generator
-license=2017 Free software to use and distributed it.
+description = Unnamed Pipe Module generator
+license=GPL 2017 - 2024 Free software to use and distributed it.
 long_description = file: README.md
 long_description_content_type = text/markdown
-keywords=unnamed pipe, code generator, skeleton
+keywords=unnamed pipe, configuration, generator, skeleton
 platforms = any
 url = https://vroncevic.github.io/gen_unnamed_pipe
 classifiers =
-    Programming Language :: Python :: 2.7
-    Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.1
-    Programming Language :: Python :: 3.2
-    Programming Language :: Python :: 3.3
-    Programming Language :: Python :: 3.4
-    Programming Language :: Python :: 3.5
-    Programming Language :: Python :: 3.6
-    Programming Language :: Python :: 3.7
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
     License :: OSI Approved :: GNU Lesser General Public License v2 (LGPLv2)
     License :: OSI Approved :: GNU Lesser General Public License v2 or later (LGPLv2+)
     License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)
@@ -50,6 +41,6 @@ classifiers =
 
 [options]
 packages = find:
-python_requires = >=2.7
+python_requires = >=3.10
 install_requires=
     ats-utilities
