@@ -1,198 +1,180 @@
 # -*- coding: UTF-8 -*-
 
 '''
- Module
-     __init__.py
- Copyright
-     Copyright (C) 2020 Vladimir Roncevic <elektron.ronca@gmail.com>
-     gen_unnamed_pipe is free software: you can redistribute it and/or modify it
-     under the terms of the GNU General Public License as published by the
-     Free Software Foundation, either version 3 of the License, or
-     (at your option) any later version.
-     gen_unnamed_pipe is distributed in the hope that it will be useful, but
-     WITHOUT ANY WARRANTY; without even the implied warranty of
-     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-     See the GNU General Public License for more details.
-     You should have received a copy of the GNU General Public License along
-     with this program. If not, see <http://www.gnu.org/licenses/>.
- Info
-     Defined class GenUnnamedPipe with attribute(s) and method(s).
-     Load a base info, create an CLI interface and run operation(s).
+Module
+    __init__.py
+Copyright
+    Copyright (C) 2017 - 2024 Vladimir Roncevic <elektron.ronca@gmail.com>
+    gen_unnamed_pipe is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+    gen_unnamed_pipe is distributed in the hope that it will be useful, but
+    WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+    See the GNU General Public License for more details.
+    You should have received a copy of the GNU General Public License along
+    with this program. If not, see <http://www.gnu.org/licenses/>.
+Info
+    Defined class GenUnnamedPipe with attribute(s) and method(s).
+    Load a base info, create a CLI interface and run operation(s).
 '''
 
 import sys
+from typing import Any, List, Dict
 from os.path import exists, dirname, realpath
+from os import getcwd
+from argparse import Namespace
 
 try:
-    from six import add_metaclass
-    from gen_unnamed_pipe.pro import UnnamedPipe
     from ats_utilities.splash import Splash
     from ats_utilities.logging import ATSLogger
     from ats_utilities.cli.cfg_cli import CfgCLI
-    from ats_utilities.cooperative import CooperativeMeta
     from ats_utilities.console_io.error import error_message
     from ats_utilities.console_io.verbose import verbose_message
     from ats_utilities.console_io.success import success_message
+    from ats_utilities.exceptions.ats_type_error import ATSTypeError
+    from ats_utilities.exceptions.ats_value_error import ATSValueError
+    from gen_unnamed_pipe.pro import UnnamedPipe
 except ImportError as ats_error_message:
-    MESSAGE = '\n{0}\n{1}\n'.format(__file__, ats_error_message)
-    sys.exit(MESSAGE)  # Force close python ATS ##############################
+    # Force close python ATS ##################################################
+    sys.exit(f'\n{__file__}\n{ats_error_message}\n')
 
 __author__ = 'Vladimir Roncevic'
-__copyright__ = 'Copyright 2020, https://vroncevic.github.io/gen_unnamed_pipe'
-__credits__ = ['Vladimir Roncevic']
+__copyright__ = '(C) 2024, https://vroncevic.github.io/gen_unnamed_pipe'
+__credits__: List[str] = ['Vladimir Roncevic', 'Python Software Foundation']
 __license__ = 'https://github.com/vroncevic/gen_unnamed_pipe/blob/dev/LICENSE'
-__version__ = '1.0.3'
+__version__ = '1.0.4'
 __maintainer__ = 'Vladimir Roncevic'
 __email__ = 'elektron.ronca@gmail.com'
 __status__ = 'Updated'
 
 
-@add_metaclass(CooperativeMeta)
 class GenUnnamedPipe(CfgCLI):
     '''
-        Defined class GenUnnamedPipe with attribute(s) and method(s).
-        Load a base info, create an CLI interface and run operation(s).
+        Defines class GenUnnamedPipe with attribute(s) and method(s).
+        Loads a base info, creates a CLI interface and run operation(s).
+
         It defines:
 
             :attributes:
-                | GEN_VERBOSE - console text indicator for process-phase.
-                | CONFIG - tool info file path.
-                | LOG - tool log file path.
-                | LOGO - logo for splash screen.
-                | OPS - list of tool options.
-                | logger - logger object API.
+                | _GEN_VERBOSE - Console text indicator for process-phase.
+                | _CONFIG - Tool info file path.
+                | _LOG - Tool log file path.
+                | _LOGO - Logo for splash screen.
+                | _OPS - List of tool options.
+                | _logger - Logger object API.
             :methods:
-                | __init__ - initial constructor.
-                | process - process and run operation.
-                | __str__ - dunder method for GenUnnamedPipe.
+                | __init__ - Initials GenUnnamedPipe constructor.
+                | process - Processes and runs tool operation.
     '''
 
-    GEN_VERBOSE = 'GEN_UNNAMED_PIPE'
-    CONFIG = '/conf/gen_unnamed_pipe.cfg'
-    LOG = '/log/gen_unnamed_pipe.log'
-    LOGO = '/conf/gen_unnamed_pipe.logo'
-    OPS = ['-g', '--gen', '-v', '--verbose', '--version']
+    _GEN_VERBOSE = 'GEN_UNNAMED_PIPE'
+    _CONFIG: str = '/conf/gen_unnamed_pipe.cfg'
+    _LOG: str = '/log/gen_unnamed_pipe.log'
+    _LOGO: str = '/conf/gen_unnamed_pipe.logo'
+    _OPS: List[str] = ['-n', '--name', '-v', '--verbose']
 
-    def __init__(self, verbose=False):
+    def __init__(self, verbose: bool = False) -> None:
         '''
-            Initial constructor.
+            Initials GenVHost constructor.
 
-            :param verbose: enable/disable verbose option.
+            :param verbose: Enable/Disable verbose option
             :type verbose: <bool>
             :exceptions: None
         '''
-        current_dir = dirname(realpath(__file__))
-        gen_unnamed_pipe_property = {
+        current_dir: str = dirname(realpath(__file__))
+        gen_unnamed_pipe_property: Dict[str, str | bool] = {
             'ats_organization': 'vroncevic',
-            'ats_repository': 'gen_unnamed_pipe',
-            'ats_name': 'gen_unnamed_pipe',
-            'ats_logo_path': '{0}{1}'.format(
-                current_dir, GenUnnamedPipe.LOGO
-            ),
+            'ats_repository': f'{self._GEN_VERBOSE.lower()}',
+            'ats_name': f'{self._GEN_VERBOSE.lower()}',
+            'ats_logo_path': f'{current_dir}{self._LOGO}',
             'ats_use_github_infrastructure': True
         }
-        splash = Splash(gen_unnamed_pipe_property, verbose=verbose)
-        base_info = '{0}{1}'.format(current_dir, GenUnnamedPipe.CONFIG)
-        CfgCLI.__init__(self, base_info, verbose=verbose)
-        verbose_message(GenUnnamedPipe.GEN_VERBOSE, verbose, 'init tool info')
-        self.logger = ATSLogger(
-            GenUnnamedPipe.GEN_VERBOSE.lower(),
-            '{0}{1}'.format(current_dir, GenUnnamedPipe.LOG),
-            verbose=verbose
+        Splash(gen_unnamed_pipe_property, verbose)
+        base_info: str = f'{current_dir}{self._CONFIG}'
+        super().__init__(base_info, verbose)
+        verbose_message(
+            verbose, [f'{self._GEN_VERBOSE.lower()} init tool info']
+        )
+        self._logger: ATSLogger = ATSLogger(
+            self._GEN_VERBOSE.lower(), f'{current_dir}{self._LOG}', verbose
         )
         if self.tool_operational:
             self.add_new_option(
-                GenUnnamedPipe.OPS[0], GenUnnamedPipe.OPS[1],
-                dest='gen', help='generate option'
+                self._OPS[0], self._OPS[1], dest='name',
+                help='generate unnamed pipe skeleton (provide name)'
             )
             self.add_new_option(
-                GenUnnamedPipe.OPS[2], GenUnnamedPipe.OPS[3],
+                self._OPS[2], self._OPS[3],
                 action='store_true', default=False,
                 help='activate verbose mode for generation'
             )
-            self.add_new_option(
-                GenUnnamedPipe.OPS[4], action='version', version=__version__
-            )
 
-    def process(self, verbose=False):
+    def process(self, verbose: bool = False) -> bool:
         '''
-            Process and run operation.
+            Processes and runs tool operation.
 
-            :param verbose: enable/disable verbose option.
+            :param verbose: Enable/Disable verbose option
             :type verbose: <bool>
-            :return: boolean status, True (success) | False.
+            :return: True (success operation) | False
             :rtype: <bool>
             :exceptions: None
         '''
-        status = False
+        status: bool = False
         if self.tool_operational:
-            num_of_args_sys = len(sys.argv)
-            if num_of_args_sys > 1:
-                operation = sys.argv[1]
-                if operation not in GenUnnamedPipe.OPS:
-                    sys.argv.append('-h')
-            else:
-                sys.argv.append('-h')
-            args = self.parse_args(sys.argv[1:])
-            pro_exists = exists(getattr(args, 'gen'))
-            if not pro_exists:
-                if bool(getattr(args, 'gen')):
-                    print(
-                        '{0} {1} [{2}]'.format(
-                            '[{0}]'.format(GenUnnamedPipe.GEN_VERBOSE.lower()),
-                            'generating', getattr(args, 'gen')
-                        )
-                    )
-                    generator = UnnamedPipe(verbose=verbose)
-                    status = generator.gen_setup(
-                        getattr(args, 'gen'), verbose
-                    )
-                    if status:
-                        success_message(GenUnnamedPipe.GEN_VERBOSE, 'done\n')
-                        self.logger.write_log(
-                            '{0} {1} done'.format(
-                                'generating tool/gen',
-                                getattr(args, 'gen')
-                            ), ATSLogger.ATS_INFO
-                        )
-                    else:
-                        error_message(
-                            GenUnnamedPipe.GEN_VERBOSE, 'generation failed'
-                        )
-                        self.logger.write_log(
-                            'generation failed', ATSLogger.ATS_ERROR
-                        )
-                else:
+            try:
+                args: Any | Namespace = self.parse_args(sys.argv)
+                if not bool(getattr(args, "name")):
                     error_message(
-                        GenUnnamedPipe.GEN_VERBOSE, 'provide project name'
+                        [f'{self._GEN_VERBOSE.lower()} missing name argument']
                     )
-                    self.logger.write_log(
-                        'provide project name', ATSLogger.ATS_ERROR
+                    return status
+                if exists(f'{getcwd()}/{str(getattr(args, "name"))}'):
+                    error_message([
+                        f'{self._GEN_VERBOSE.lower()}',
+                        f'project with name [{getattr(args, "name")}] exists'
+                    ])
+                    return status
+                gen: UnnamedPipe = UnnamedPipe(
+                    getattr(args, 'verbose') or verbose
+                )
+                try:
+                    print(
+                        " ".join([
+                            f'[{self._GEN_VERBOSE.lower()}]',
+                            'generate unnamed pipe skeleton',
+                            str(getattr(args, 'name'))
+                        ])
                     )
-            else:
+                    status: bool = gen.gen_setup(
+                        str(getattr(args, "name")),
+                        getattr(args, 'verbose') or verbose
+                    )
+                except (ATSTypeError, ATSValueError) as e:
+                    error_message([f'{self._GEN_VERBOSE.lower()} {str(e)}'])
+                    self._logger.write_log(f'{str(e)}', self._logger.ATS_ERROR)
+                if status:
+                    success_message([f'{self._GEN_VERBOSE.lower()} done\n'])
+                    self._logger.write_log(
+                        f'generation {getattr(args, "name")} done',
+                        self._logger.ATS_INFO
+                    )
+                else:
+                    error_message([f'{self._GEN_VERBOSE.lower()} failed'])
+                    self._logger.write_log(
+                        'generation failed', self._logger.ATS_ERROR
+                    )
+            except SystemExit:
                 error_message(
-                    GenUnnamedPipe.GEN_VERBOSE, 'project already exist'
+                    [f'{self._GEN_VERBOSE.lower()} expected argument -n']
                 )
-                self.logger.write_log(
-                    'project already exist', ATSLogger.ATS_ERROR
-                )
+                return status
         else:
             error_message(
-                GenUnnamedPipe.GEN_VERBOSE, 'tool is not operational'
+                [f'{self._GEN_VERBOSE.lower()} tool is not operational']
             )
-            self.logger.write_log(
-                'tool is not operational', ATSLogger.ATS_ERROR
+            self._logger.write_log(
+                'tool is not operational', self._logger.ATS_ERROR
             )
         return status
-
-    def __str__(self):
-        '''
-            Dunder method for GenUnnamedPipe.
-
-            :return: object in a human-readable format.
-            :rtype: <str>
-            :exceptions: None
-        '''
-        return '{0} ({1}, {2})'.format(
-            self.__class__.__name__, CfgCLI.__str__(self), str(self.logger)
-        )
