@@ -1,227 +1,155 @@
 # -*- coding: UTF-8 -*-
 
 '''
- Module
-     __init__.py
- Copyright
-     Copyright (C) 2020 Vladimir Roncevic <elektron.ronca@gmail.com>
-     gen_unnamed_pipe is free software: you can redistribute it and/or modify
-     it under the terms of the GNU General Public License as published by the
-     Free Software Foundation, either version 3 of the License, or
-     (at your option) any later version.
-     gen_unnamed_pipe is distributed in the hope that it will be useful, but
-     WITHOUT ANY WARRANTY; without even the implied warranty of
-     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-     See the GNU General Public License for more details.
-     You should have received a copy of the GNU General Public License along
-     with this program. If not, see <http://www.gnu.org/licenses/>.
- Info
-     Defined class UnnamedPipe with attribute(s) and method(s).
-     Generate module file generator_test.py by template and parameters.
+Module
+    __init__.py
+Copyright
+    Copyright (C) 2017 - 2024 Vladimir Roncevic <elektron.ronca@gmail.com>
+    gen_unnamed_pipe is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+    gen_unnamed_pipe is distributed in the hope that it will be useful, but
+    WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+    See the GNU General Public License for more details.
+    You should have received a copy of the GNU General Public License along
+    with this program. If not, see <http://www.gnu.org/licenses/>.
+Info
+    Defines class VHost with attribute(s) and method(s).
+    Generates virtual host configuration module by template and parameters.
 '''
 
 import sys
+from typing import List, Dict
 from os.path import dirname, realpath
 
 try:
-    from gen_unnamed_pipe.pro.config import ProConfig
-    from gen_unnamed_pipe.pro.config.pro_name import ProName
-    from gen_unnamed_pipe.pro.read_template import ReadTemplate
-    from gen_unnamed_pipe.pro.write_template import WriteTemplate
-    from ats_utilities.checker import ATSChecker
-    from ats_utilities.config_io.base_check import FileChecking
-    from ats_utilities.console_io.error import error_message
+    from ats_utilities.pro_config import ProConfig
+    from ats_utilities.pro_config.pro_name import ProName
+    from ats_utilities.config_io.file_check import FileCheck
     from ats_utilities.console_io.verbose import verbose_message
     from ats_utilities.config_io.yaml.yaml2object import Yaml2Object
     from ats_utilities.exceptions.ats_type_error import ATSTypeError
-    from ats_utilities.exceptions.ats_bad_call_error import ATSBadCallError
+    from ats_utilities.exceptions.ats_value_error import ATSValueError
+    from gen_unnamed_pipe.pro.read_template import ReadTemplate
+    from gen_unnamed_pipe.pro.write_template import WriteTemplate
 except ImportError as ats_error_message:
-    MESSAGE = '\n{0}\n{1}\n'.format(__file__, ats_error_message)
-    sys.exit(MESSAGE)  # Force close python ATS ##############################
+    # Force close python ATS ##################################################
+    sys.exit(f'\n{__file__}\n{ats_error_message}\n')
 
 __author__ = 'Vladimir Roncevic'
-__copyright__ = 'Copyright 2020, https://vroncevic.github.io/gen_unnamed_pipe'
-__credits__ = ['Vladimir Roncevic']
+__copyright__ = '(C) 2024, https://vroncevic.github.io/gen_unnamed_pipe'
+__credits__: List[str] = ['Vladimir Roncevic', 'Python Software Foundation']
 __license__ = 'https://github.com/vroncevic/gen_unnamed_pipe/blob/dev/LICENSE'
-__version__ = '1.0.3'
+__version__ = '1.0.4'
 __maintainer__ = 'Vladimir Roncevic'
 __email__ = 'elektron.ronca@gmail.com'
 __status__ = 'Updated'
 
 
-class UnnamedPipe(FileChecking, ProConfig, ProName):
+class UnnamedPipe(FileCheck, ProConfig, ProName):
     '''
-        Defined class UnnamedPipe with attribute(s) and method(s).
-        Generate module file generator_test.py by template and parameters.
+        Defines class VHost with attribute(s) and method(s).
+        Generates virtual host configuration module by template and parameters.
+
         It defines:
 
             :attributes:
-                | GEN_VERBOSE - console text indicator for process-phase.
-                | PRO_STRUCTURE - project setup (template, module).
-                | __reader - reader API.
-                | __writer - writer API.
+                | _GEN_VERBOSE - Console text indicator for process-phase.
+                | _PRO_STRUCTURE - Project setup (templates, modules).
+                | _reader - Reader API.
+                | _writer - Writer API.
             :methods:
-                | __init__ - initial constructor.
-                | get_reader - getter for template reader.
-                | get_writer - getter for template writer.
-                | gen_setup - generate module file setup.py.
-                | __str__ - dunder method for UnnamedPipe.
+                | __init__ - Initials VHost constructor.
+                | get_reader - Gets template reader.
+                | get_writer - Gets template writer.
+                | gen_model - Generates data model.
     '''
 
-    GEN_VERBOSE = 'GEN_UNNAMED_PIPE::PRO::GEN_SETUP'
-    PRO_STRUCTURE = '/../conf/project.yaml'
+    _GEN_VERBOSE: str = 'GEN_UNNAMED_PIPE::PRO::VHOST'
+    _PRO_STRUCTURE: str = '/../conf/project.yaml'
 
-    def __init__(self, verbose=False):
+    def __init__(self, verbose: bool = False) -> None:
         '''
             Initial constructor.
 
-            :param verbose: enable/disable verbose option.
+            :param verbose: Enable/Disable verbose option
             :type verbose: <bool>
             :exceptions: None
         '''
-        FileChecking.__init__(self, verbose=verbose)
-        ProConfig.__init__(self, verbose=verbose)
-        ProName.__init__(self, verbose=verbose)
+        FileCheck.__init__(self, verbose)
+        ProConfig.__init__(self, verbose)
+        ProName.__init__(self, verbose)
         verbose_message(
-            UnnamedPipe.GEN_VERBOSE, verbose, 'init generator'
+            verbose, [f'{self._GEN_VERBOSE.lower()} init generator']
         )
-        self.__reader = ReadTemplate(verbose=verbose)
-        self.__writer = WriteTemplate(verbose=verbose)
-        project_structure = '{0}{1}'.format(
-            dirname(realpath(__file__)), UnnamedPipe.PRO_STRUCTURE
-        )
-        self.check_path(file_path=project_structure, verbose=verbose)
-        self.check_mode(file_mode='r', verbose=verbose)
-        self.check_format(
-            file_path=project_structure, file_format='yaml', verbose=verbose
-        )
+        self._reader: ReadTemplate | None = ReadTemplate(verbose)
+        self._writer: WriteTemplate | None = WriteTemplate(verbose)
+        current_dir: str = dirname(realpath(__file__))
+        pro_structure: str = f'{current_dir}{self._PRO_STRUCTURE}'
+        self.check_path(pro_structure, verbose)
+        self.check_mode('r', verbose)
+        self.check_format(pro_structure, 'yaml', verbose)
         if self.is_file_ok():
-            yml2obj = Yaml2Object(project_structure)
+            yml2obj: Yaml2Object | None = Yaml2Object(pro_structure)
             self.config = yml2obj.read_configuration()
 
-    def get_reader(self):
+    def get_reader(self) -> ReadTemplate | None:
         '''
-            Getter for template reader.
+            Gets template reader.
 
-            :return: template reader object.
-            :rtype: <ReadTemplate>
+            :return: Template reader object | None
+            :rtype: <ReadTemplate> | <NoneType>
             :exceptions: None
         '''
-        return self.__reader
+        return self._reader
 
-    def get_writer(self):
+    def get_writer(self) -> WriteTemplate | None:
         '''
-            Getter for template writer.
+            Gets template writer.
 
-            :return: template writer object.
-            :rtype: <WriteTemplate>
+            :return: Template writer object | none
+            :rtype: <WriteTemplate> | <NoneType
             :exceptions: None
         '''
-        return self.__writer
+        return self._writer
 
-    def select_pro_type(self, verbose=False):
+    def gen_setup(
+        self,
+        module_name: str | None,
+        verbose: bool = False
+    ) -> bool:
         '''
-            Select project type.
+            Generates virtual host module.
 
-            :param verbose: enable/disable verbose option.
+            :param module_name: Parameter module name
+            :type module_name: <str> | <NoneType>
+            :param verbose: Enable/Disable verbose option
             :type verbose: <bool>
-            :return: template type | None.
-            :rtype: <str> | <NoneType>
-            :exceptions: None
-        '''
-        template_selected = None
-        if bool(self.config):
-            types = self.config['templates']
-            pro_types_len = len(types)
-            for index, pro_type in enumerate(types):
-                for project_type, template_file in pro_type.items():
-                    if project_type == 'cancel':
-                        print(
-                            '{0} {1}'.format(
-                                index + 1, project_type.capitalize()
-                            )
-                        )
-                    else:
-                        print(
-                            '{0} {1}'.format(
-                                index + 1,
-                                project_type.upper().replace('_', ' ')
-                            )
-                        )
-                    verbose_message(
-                        UnnamedPipe.GEN_VERBOSE, verbose,
-                        'to be processed template', template_file
-                    )
-            while True:
-                input_type = input(' select project type: ')
-                options = range(1, pro_types_len + 1, 1)
-                try:
-                    if int(input_type) in list(options):
-                        for target in types[int(input_type) - 1].keys():
-                            if target is None:
-                                template_selected = 'cancel'
-                            else:
-                                template_selected = target
-                        break
-                    else:
-                        raise ValueError
-                except ValueError:
-                    error_message(
-                        UnnamedPipe.GEN_VERBOSE, 'not an appropriate choice'
-                    )
-            verbose_message(
-                UnnamedPipe.GEN_VERBOSE, verbose,
-                'selected', template_selected
-            )
-        return template_selected
-
-    def gen_setup(self, pro_name, verbose=False):
-        '''
-            Generate unnamed pipe modules.
-
-            :param pro_name: project name.
-            :type pro_name: <str>
-            :param verbose: enable/disable verbose option.
-            :type verbose: <bool>
-            :return: boolean status, True (success) | False.
+            :return: True (success operation) | False
             :rtype: <bool>
-            :exceptions: ATSTypeError | ATSBadCallError
+            :exceptions: ATSTypeError | ATSValueError
         '''
-        checker, error, status = ATSChecker(), None, False
-        error, status = checker.check_params([('str:pro_name', pro_name)])
-        if status == ATSChecker.TYPE_ERROR:
-            raise ATSTypeError(error)
-        if status == ATSChecker.VALUE_ERROR:
-            raise ATSBadCallError(error)
-        status = False
+        error_msg: str | None = None
+        error_id: int | None = None
+        error_msg, error_id = self.check_params([
+            ('str:model_name', module_name)
+        ])
+        if error_id == self.TYPE_ERROR:
+            raise ATSTypeError(error_msg)
+        if not bool(module_name):
+            raise ATSValueError('missing module name')
         verbose_message(
-            UnnamedPipe.GEN_VERBOSE, verbose,
-            'generating unnamed pipe modules for', pro_name
+            verbose, [f'{self._GEN_VERBOSE.lower()} generate {module_name}']
         )
-        if bool(self.config):
-            uds_type = self.select_pro_type(verbose=verbose)
-            if uds_type != 'cancel':
-                templates = self.__reader.read(
-                    self.config, uds_type, verbose=verbose
+        status: bool = False
+        if bool(self._reader) and bool(self.config):
+            template_content: Dict[str, str] = self._reader.read(
+                self.config, verbose
+            )
+            if bool(template_content) and bool(self._writer):
+                status = self._writer.write(
+                    template_content, module_name, verbose
                 )
-                if bool(templates):
-                    status = self.__writer.write(
-                        templates, pro_name, verbose=verbose
-                    )
-            else:
-                status = True
         return status
-
-    def __str__(self):
-        '''
-            Dunder method for UnnamedPipe.
-
-            :return: object in a human-readable format.
-            :rtype: <str>
-            :exceptions: None
-        '''
-        return '{0} ({1}, {2}, {3}, {4}, {5})'.format(
-            self.__class__.__name__, FileChecking.__str__(self),
-            ProConfig.__str__(self), ProName.__str__(self),
-            str(self.__reader), str(self.__writer)
-        )
