@@ -1,136 +1,129 @@
 # -*- coding: UTF-8 -*-
 
 '''
- Module
-     write_template.py
- Copyright
-     Copyright (C) 2020 Vladimir Roncevic <elektron.ronca@gmail.com>
-     gen_unnamed_pipe is free software: you can redistribute it and/or modify it
-     under the terms of the GNU General Public License as published by the
-     Free Software Foundation, either version 3 of the License, or
-     (at your option) any later version.
-     gen_unnamed_pipe is distributed in the hope that it will be useful, but
-     WITHOUT ANY WARRANTY; without even the implied warranty of
-     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-     See the GNU General Public License for more details.
-     You should have received a copy of the GNU General Public License along
-     with this program. If not, see <http://www.gnu.org/licenses/>.
- Info
-     Defined class WriteTemplate with attribute(s) and method(s).
-     Created API for write operation of template content.
+Module
+    write_template.py
+Copyright
+    Copyright (C) 2017 - 2024 Vladimir Roncevic <elektron.ronca@gmail.com>
+    gen_unnamed_pipe is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+    gen_unnamed_pipe is distributed in the hope that it will be useful, but
+    WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+    See the GNU General Public License for more details.
+    You should have received a copy of the GNU General Public License along
+    with this program. If not, see <http://www.gnu.org/licenses/>.
+Info
+    Defines class WriteTemplate with atribute(s) and method(s).
+    Creates an API for write a template content with parameters to a file.
 '''
 
 import sys
-from os import getcwd, chmod
+from typing import List, Dict
+from os import getcwd, chmod, mkdir
 from string import Template
-from datetime import datetime
+from datetime import date
 
 try:
-    from ats_utilities.checker import ATSChecker
-    from ats_utilities.config_io.base_check import FileChecking
+    from ats_utilities.config_io.file_check import FileCheck
     from ats_utilities.console_io.verbose import verbose_message
     from ats_utilities.exceptions.ats_type_error import ATSTypeError
-    from ats_utilities.exceptions.ats_bad_call_error import ATSBadCallError
+    from ats_utilities.exceptions.ats_value_error import ATSValueError
 except ImportError as ats_error_message:
-    MESSAGE = '\n{0}\n{1}\n'.format(__file__, ats_error_message)
-    sys.exit(MESSAGE)  # Force close python ATS ##############################
+    # Force close python ATS ##################################################
+    sys.exit(f'\n{__file__}\n{ats_error_message}\n')
 
 __author__ = 'Vladimir Roncevic'
-__copyright__ = 'Copyright 2020, https://vroncevic.github.io/gen_unnamed_pipe'
-__credits__ = ['Vladimir Roncevic']
+__copyright__ = '(C) 2024, https://vroncevic.github.io/gen_unnamed_pipe'
+__credits__: List[str] = ['Vladimir Roncevic', 'Python Software Foundation']
 __license__ = 'https://github.com/vroncevic/gen_unnamed_pipe/blob/dev/LICENSE'
-__version__ = '1.0.3'
+__version__ = '1.0.4'
 __maintainer__ = 'Vladimir Roncevic'
 __email__ = 'elektron.ronca@gmail.com'
 __status__ = 'Updated'
 
 
-class WriteTemplate(FileChecking):
+class WriteTemplate(FileCheck):
     '''
-        Defined class WriteTemplate with attribute(s) and method(s).
-        Created API for write operation of template content.
+        Defines class WriteTemplate with attribute(s) and method(s).
+        Creates an API for write a template content with parameters to a file.
+
         It defines:
 
             :attributes:
-                | GEN_VERBOSE - console text indicator for process-phase.
+                | _GEN_VERBOSE - Console text indicator for process-phase.
             :methods:
-                | __init__ - initial constructor.
-                | get_setup - getter for setup file object.
-                | write - write setup content to files.
-                | __str__ - dunder method for WriteTemplate.
+                | __init__ - Initials WriteTemplate constructor.
+                | write - write a template content with parameters to a file.
     '''
 
-    GEN_VERBOSE = 'GEN_UNNAMED_PIPE::PRO::WRITE_TEMPLATE'
+    _GEN_VERBOSE: str = 'GEN_UNNAMED_PIPE::PRO::WRITE_TEMPLATE'
 
-    def __init__(self, verbose=False):
+    def __init__(self, verbose: bool = False) -> None:
         '''
-            Initial constructor.
+            Initials WriteTemplate constructor.
 
-            :param verbose: enable/disable verbose option.
+            :param verbose: Enable/Disable verbose option
             :type verbose: <bool>
-            :exceptions: None
+            :excptions: None
         '''
-        FileChecking.__init__(self, verbose=verbose)
-        verbose_message(WriteTemplate.GEN_VERBOSE, verbose, 'init writer')
+        super().__init__(verbose)
+        verbose_message(verbose, [f'{self._GEN_VERBOSE.lower()} init writer'])
 
-    def write(self, templates, pro_name, verbose=False):
+    def write(
+        self,
+        template_content: Dict[str, str],
+        pro_name: str | None,
+        verbose: bool = False
+    ) -> bool:
         '''
-            Write setup content to files.
+            Write setup content to file.
 
-            :param templates: templates and contents.
-            :type templates: <list>
-            :param pro_name: project name.
-            :type pro_name: <str>
-            :param verbose: enable/disable verbose option.
+            :param template_content: Template content
+            :type template_content: <Dict[str, str]>
+            :param pro_name: Project name | None
+            :type pro_name: <str> | <NoneType>
+            :param verbose: Enable/Disable verbose option
             :type verbose: <bool>
-            :return: boolean status, True (success) | False.
+            :return: True (success operation) | False
             :rtype: <bool>
-            :exception: ATSTypeError | ATSBadCallError
+            :exception: ATSTypeError | ATSValueError
         '''
-        checker, error, status = ATSChecker(), None, False
-        error, status = checker.check_params([
-            ('list:templates', templates), ('str:pro_name', pro_name)
+        error_msg: str | None = None
+        error_id: int | None = None
+        error_msg, error_id = self.check_params([
+            ('dict:template_content', template_content),
+            ('str:pro_name', pro_name)
         ])
-        if status == ATSChecker.TYPE_ERROR:
-            raise ATSTypeError(error)
-        if status == ATSChecker.VALUE_ERROR:
-            raise ATSBadCallError(error)
-        status, template = False, None
-        pro_setup = {
-            'PRO': '{0}'.format(pro_name),
-            'YEAR': '{0}'.format(datetime.now().year)
-        }
-        for template_module in templates:
-            template = Template(template_module.values()[0])
-            setup_module = '{0}/{1}'.format(
-                getcwd(), template_module.keys()[0]
+        if error_id == self.TYPE_ERROR:
+            raise ATSTypeError(error_msg)
+        if not bool(template_content):
+            raise ATSValueError('missing readme content')
+        if not bool(pro_name):
+            raise ATSValueError('missing project name')
+        module_pro_dir: str = f'{getcwd()}/{pro_name}/'
+        mkdir(module_pro_dir)
+        status: bool = False
+        for module_name, module_content in template_content.items():
+            template_module: Template = Template(
+                f'{module_pro_dir}{module_name}'
             )
-            if template:
-                with open(setup_module, 'w') as setup_file:
-                    verbose_message(
-                        WriteTemplate.GEN_VERBOSE, verbose,
-                        'write', setup_module
-                    )
-                    setup_file.write(template.substitute(pro_setup))
-                    chmod(setup_module, 0o666)
-                    self.check_path(setup_module, verbose=verbose)
-                    self.check_mode('w', verbose=verbose)
-                    self.check_format(
-                        setup_module, setup_module.split('.')[1],
-                        verbose=verbose
-                    )
-                    if self.is_file_ok():
-                        status = True
+            module_path: str = template_module.substitute({
+                'PRO_NAME': pro_name
+            })
+            template: Template = Template(module_content)
+            with open(module_path, 'w', encoding='utf-8') as module_file:
+                module_file.write(template.substitute({
+                    'PRO': pro_name, 'YEAR': f'{str(date.today().year)}'
+                }))
+                chmod(module_path, 0o644)
+                self.check_path(module_path, verbose)
+                self.check_mode('w', verbose)
+                self.check_format(
+                    module_path, module_path.split('.')[1], verbose
+                )
+                if self.is_file_ok():
+                    status = True
         return status
-
-    def __str__(self):
-        '''
-            Dunder method for WriteTemplate.
-
-            :return: object in a human-readable format.
-            :rtype: <str>
-            :exceptions: None
-        '''
-        return '{0} ({1})'.format(
-            self.__class__.__name__, FileChecking.__str__(self)
-        )
